+++ Dyna/Features/Bastion.cs
@@ -14,7 +14,8 @@ internal sealed class BastionManager : IDynaHook, IStatusRenderHook
 			Definition = new()
 			{
 				icon = ModEntry.Instance.Helper.Content.Sprites.RegisterSprite(ModEntry.Instance.Package.PackageRoot.GetRelativeFile("assets/Status/Bastion.png")).Sprite,
-				color = new("7E7E7E")
+				color = new("7E7E7E"),
+				isGood = true,
 			},
 			Name = ModEntry.Instance.AnyLocalizations.Bind(["status", "Bastion", "name"]).Localize,
 			Description = ModEntry.Instance.AnyLocalizations.Bind(["status", "Bastion", "description"]).Localize